Fix bug 6079 - Autorun halted at region load zones
The old code called MovePlayerTo() with the new sector and velocity
whether or not the loading screen was being displayed and before the
loading of the new sector was finished. Also, it set the velocity in the
dummy loading sector to the current velocity which is probably not zero.
Now the velocity in the loading sectory is always set to zero, and it only
restores the current velocity in the new sector after the sector is loaded
or immediately if *not* loading the new sector in the background.