Fix for 5742 - Crash when opening containers
The problem was that 'worldContainer' is the source container pointer
but is then overwritten to be the destination container pointer.
This means the 'itemProposed' is removed from the destination container
instead of the source container leading to dangling pointers.
The solution is to use different names for the destination information.