List KA triggers: Deleting a knowledge area will no longer result in a "no such area" message after deleting. Areas without triggers (npc_knowledge_areas without corresponding triggers basically) are now also shown, and can, if desired, be deleted.
KA details:
"the following npcs use this KA" now also lists the last name of NPCs, in addition, it will list the name of the NPC in case of a personal KA.
"create new trigger" now also shows when a KA has no entries (triggers).
Removed the "create sub-trigger" functionality (appears to be broken, and unused, use quests or ka-scipts instead to chain responses). 
Changed the form to have 1 "save changes" button for both the trigger and the response content.
Added a javascript to add elements to the script while staying on the page.